LocalizacaoIndoorAPI/src/main/java/cesar/com/localizacaoindoorapi/model/Location.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package cesar.com.localizacaoindoorapi.model;

import java.util.List;
import java.util.Objects;

/**
 *
 * @author cesar
 */
public class Location {

    private String name;
    private List<Wifi> referencesAp;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public List<Wifi> getReferencesAp() {
        return referencesAp;
    }

    public void setReferencesAp(List<Wifi> referencesAp) {
        this.referencesAp = referencesAp;
    }

    @Override
    public String toString() {
        return this.getName();
    }

    @Override
    public int hashCode() {
        int hash = 7;
        hash = 37 * hash + Objects.hashCode(this.name);
        return hash;
    }

    @Override
    public boolean equals(Object obj) {
        if (this == obj) {
            return true;
        }
        if (obj == null) {
            return false;
        }
        if (getClass() != obj.getClass()) {
            return false;
        }
        final Location other = (Location) obj;
        if (!Objects.equals(this.name, other.name)) {
            return false;
        }
        return true;
    }

}
